split wasmer2 test
