+++ integration-tests/src/tests/client/features.rs
@@ -8,3 +8,4 @@ mod fix_contract_loading_cost;
 mod fix_storage_usage;
 mod lower_storage_key_limit;
 mod restore_receipts_after_fix_apply_chunks;
+mod wasmer2;
